seperate out report generation in models
